+++ gatsby-browser.js
@@ -8,6 +8,8 @@ import { ColorSchemeProvider } from "components/ColorSchemeProvider"
 
 import { Facebook } from "components/Facebook"
 
+import "prismjs/plugins/line-numbers/prism-line-numbers.css"
+
 // export const wrapRootElement vs. exports.wrapRootElement...
 export const wrapRootElement = ({ element }) => {
   return (