+++ src/components/Comments/Overflow/deleteCommentById.ts
@@ -1,4 +1,4 @@
-import { gql } from "apollo-boost"
+import { gql } from "@apollo/client"
 
 export const DELETE_COMMENT_BY_ID = gql`
   mutation DeleteCommentById($id: ID!) {