+++ src/components/layout.tsx
@@ -4,15 +4,13 @@ import { useSpring, animated } from "react-spring"
 import { compose } from "redux"
 import styled from "styled-components"
 
-import { ButtonAndDrawer, Footer, Header, NavBar } from "components"
+import { Footer, Header, NavBar } from "components"
 import { StickyNumbers } from "components/StickyNumbers"
 import { rhythm } from "utils/typography"
 import { Colors } from "consts/Colors"
 import { AnimatedDottedBackground } from "components/AnimatedDottedBackground"
 
 import { useWindowScrollPercent } from "hooks/useWindowScrollPercent"
-
-import "prismjs/plugins/line-numbers/prism-line-numbers.css"
 
 const ThemedBackground = styled(animated.div)`
   position: fixed;
@@ -37,14 +35,12 @@ const LayoutFrame = styled(animated.div)`
   /* border: ${process.env.NODE_ENV === "development" && `1px dotted red`}; */
 `
 const Main = styled(animated.main)`
-  /* border: ${
-    process.env.NODE_ENV === "development" && `1px dotted orange`
-  }; */
+  /* border: ${process.env.NODE_ENV === "development" &&
+  `1px dotted orange`}; */
 `
 const Inner = styled(animated.div)`
-  /* border: ${
-    process.env.NODE_ENV === "development" && `1px dotted yellow`
-  }; */
+  /* border: ${process.env.NODE_ENV === "development" &&
+  `1px dotted yellow`}; */
 `
 
 function Layout({ location, title, children }: Props) {
@@ -128,7 +124,6 @@ function Layout({ location, title, children }: Props) {
             <Header location={location} title={title} />
           </animated.header>
 
-          <ButtonAndDrawer />
           <Main>{children}</Main>
           <Footer />
         </Inner>