+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v0.18.1] - 2020-11-19
+
+### Clean Up
+
+- comment out footer links 4230c71
+- drastic type improvements for tags.tsx 08d2131
+- Improve SEO types 7b998a3
+- tags.tsx improve types 1c3f0eb
+- remove `apollo-boost`, `apollo-link-context`, `apollo-link-ws` 930237d
+- delete ButtonAndDrawer b93252a
+- refix invariant ApolloClient error on SSR 7f9f28f
+- move prismjs css imports to browser.js ba902c7
+
 ## [v0.18.0] - 2020-11-19
 
 ### Notes
@@ -172,6 +185,7 @@ New:
 
 ### Added
 
+[v0.18.1]: https://github.com/thiskevinwang/coffee-code-climb/compare/v0.18.0...v0.18.1
 [v0.18.0]: https://github.com/thiskevinwang/coffee-code-climb/compare/v0.17.1...v0.18.0
 [v0.17.1]: https://github.com/thiskevinwang/coffee-code-climb/compare/v0.17.0...v0.17.1
 [v0.17.0]: https://github.com/thiskevinwang/coffee-code-climb/compare/v0.16.1...v0.17.0